api/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-dotenv==0.20.0
sqlalchemy[asyncio]==1.4.36
alembic==1.7.7
asyncpg==0.25.0
pytest-asyncio==0.18.3
pytest-env==0.6.2
httpx==0.22.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
